+++ src/Validator.ts
@@ -7,6 +7,7 @@ import ValidatorInterface, {
   ValidationPlugins,
 } from "./models/ValidatorInterface";
 import FormInterface from "./models/FormInterface";
+import FieldInterface from "./models/FieldInterface";
 import { OptionsEnum } from "./models/OptionsModel";
 
 export default class Validator implements ValidatorInterface {
@@ -55,10 +56,9 @@ export default class Validator implements ValidatorInterface {
 
   validate(opt: any = {}, obj: any = {}): Promise<any> {
     const path = $try(opt.path, opt);
-    const field = $try(opt.field, this.form.select(path, null, null));
+    const instance = $try(opt.field, this.form.select(path, null, null), this.form);
     const related = $try(opt.related, obj.related, true);
     const showErrors = $try(opt.showErrors, obj.showErrors, false);
-    const instance = field || this.form;
     instance.$validating = true;
     instance.$validated += 1;
 
@@ -113,47 +113,37 @@ export default class Validator implements ValidatorInterface {
     path,
   }: any): void {
     const instance = field || this.form.select(path);
+    const { options } = this.form.state;
     // check if the field is a valid instance
-    if (!instance.path)
-      throw new Error("Validation Error: Invalid Field Instance");
+    if (!instance.path) throw new Error("Validation Error: Invalid Field Instance");
     // do not validate soft deleted fields
-    if (
-      instance.deleted &&
-      !this.form.state.options.get(OptionsEnum.validateDeletedFields)
-    )
-      return;
+    if (instance.deleted && !options.get(OptionsEnum.validateDeletedFields, instance)) return;
     // do not validate disabled fields
-    if (
-      instance.disabled &&
-      !this.form.state.options.get(OptionsEnum.validateDisabledFields)
-    )
-      return;
+    if (instance.disabled && !options.get(OptionsEnum.validateDisabledFields, instance)) return;
     // do not validate pristine fields
-    if (
-      instance.isPristine &&
-      !this.form.state.options.get(OptionsEnum.validatePristineFields)
-    )
-      return;
+    if (instance.isPristine && !options.get(OptionsEnum.validatePristineFields, instance)) return;
+    // reset validation before validate
+    if (options.get(OptionsEnum.resetValidationBeforeValidate, instance)) instance.resetValidation();
+    // trim string value before validation
+    if (options.get(OptionsEnum.validateTrimmedValue, instance)) instance.trim();
 
-    // reset field validation
-    instance.resetValidation();
-
-    // validate with all enabled drivers
-    const stopOnError = instance.state.options.get(
+    // get stop on error
+    const stopOnError = options.get(
       OptionsEnum.stopValidationOnError,
       instance
     );
 
-    // get validation order
-    const validationOrder = instance.state.options.get(
-      OptionsEnum.validationOrder,
+    // get validation plugin order
+    const validationPluginOrder = options.get(
+      OptionsEnum.validationPluginsOrder,
       instance
     );
 
-    const drivers = validationOrder
-      ? validationOrder.map((n: string) => this.drivers[n])
+    const drivers = validationPluginOrder
+      ? validationPluginOrder.map((n: string) => this.drivers[n])
       : this.drivers;
 
+    // validate with all enabled drivers
     _.each(drivers, (driver: ValidationPluginInterface) => {
       driver && driver.validate(instance);
       if (stopOnError && instance.hasError) {
@@ -172,7 +162,7 @@ export default class Validator implements ValidatorInterface {
     Validate 'related' fields if specified
     and related validation allowed (recursive)
   */
-  validateRelatedFields(field: any, showErrors: boolean): void {
+  validateRelatedFields(field: FieldInterface, showErrors: boolean): void {
     if (!field.related || !field.related.length) return;
 
     _.each(field.related, (path) =>