+++ tests/data/forms/flat/form.q.ts
@@ -7,7 +7,10 @@ const fields = [{
 }, {
   name: 'email',
   label: 'Email',
-  value: 's.jobs@apple.com',
+  value: ' s.jobs@apple.com ', // should be trimmed on initial validation and pass validation
+  options: {
+    validateTrimmedValue: true // <---
+  }
 }];
 
 export default new Form({ fields }, { name: 'Flat-Q' });